requirements.txt
numpy~=1.26.4
scikit-image~=0.23.2
matplotlib~=3.9.0
torch~=2.2.0+cu121
Pillow~=10.2.0
torchvision~=0.17.0+cu121
requests~=2.28.1

opencv-python~=4.9.0.80
tqdm~=4.66.4
scikit-learn~=1.5.0
imutils~=0.5.4